ublksrv: add flag of UBLK_F_HAS_IO_DAEMON
Add UBLK_F_HAS_IO_DAEMON which has to be defined in ublk_cmd.h as
part of ABI since  ublk control task is usually stateless, so we have
to retrieve this info from driver side.

If this flag isn't set, the io handler task is supposed to run in
the control task.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>