libublksrv: move create/remove pid file into libublksrv
ublksrv is supposed to be run in daemon context, so it is reasonable
to create pid file per traditional daemon style.

However, if target implementation won't like to create the pid file,
it is fine to not do it by passing NULL ->run_dir.

Another change is to set default ublksrv running dir as
/var/run/ublksrv/, and each pid file will be created under that dir.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>