ublksrv: Add kernel version and g++ options check
This patch updates configure.ac to add
kernel version checks and g++ checks
mandatory. This is because if the criteria
is not met the build will fail during actual
make.

This make minimum kernel version as 5.19.X.
Anything lesser than 5.19 the build will stop
with failure.

This patch also confirms if g++/c++ supports
CFLAGS option of "-fcoroutines -std=c++20".
If not supported the build stops with failure.

Signed-off-by: H Arshad <harshad1024@proton.me>