libublksrv: set IORING_SETUP_COOP_TASKRUN for ublksrv's io_uring
Per liburing's man page:

```
The exception are setups where the application uses multiple threads
operating on the same ring, where the application waiting on completions
isn't the one that submitted them.
```

For ublksrv, the single uring is used, and single pthread is for
handling submission and completion, so it is perfectly allowed
to enable IORING_SETUP_COOP_TASKRUN for avoiding extra SMP IPI
interrupt.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>