Fix printf-style format strings
Fixes warnings and mismatches in printf format strings:

- Adds attribute annotations to ublksrv_{log,printf} to get more
validation for their invocations.

- Addresses warnings for u64s in qcow code; from my read, __u64 ends
up an unsigned long long but u64 (in qcow*.cpp) ends up a uint64_t,
which should be paired with PRIx64 and PRIu64 format specifier
constants (and on my platform ends up not being an unsigned long long).

- Adds %zd/%zu in a few places where size_t values are formatted.

Signed-off-by: reuben olinsky <reubeno.dev@gmail.com>
Signed-off-by: Ming Lei <tom.leiming@gmail.com>