Handle interrupts in __ublksrv_ctrl_cmd
The syscall io_uring_enter may return EINTR if it is interrupted by a
delivery of a signal. Since __ublksrv_ctrl_cmd already added an SQE, it's
easier for ublksrv to handle the error instead of the caller.

For example, if ublksrv_ctrl_start_dev() returns -EINTR, then the caller
has to either manually wait for the CQE or might call
ublksrv_ctrl_start_dev() again, resulting in an erroneous second SQE.
(Which may have no effect, but handling the interrupt seems the most
correct)