lib: Actually return ulong in ublksrv_json_read_target_ulong_info()
The aim of ublksrv_json_read_target_ulong_info() is to fetch
value corresponding to given key from JSON and return it in its
last argument. But the last argument is of type long which
contradicts not only the function name but also set function
(ublksrv_json_write_target_ulong_info()). Switch the argument
type to unsigned long.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>