Update the execve helper to always run targets from the same directory
In the execve helper, ensure that when we spawn a target binary that
we always run it as an absolute path and that the target is in the
same binary as the ublk helper itself.

For example if the ublk binary is in
/home/sahlberg/ublksrv/.libs/ublk
then for type "loop" the binary we will execve will be
/home/sahlberg/ublksrv/.libs/ublk.loop

Additionally, copy the environment variable LD_LIBRARY_PATH from
the ublk binary to the environment for the target binary.
In a local build environment, when you call the binary directly
without going through libtool the spwaned binary might not find
the libublksrv.so library.

In test systems where you have not installed ublksrv you may or may not
need to set LD_LIBRARY_PATH=`pwd`/lib/.libs

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>