README: remove todo list and add maillist entry
Remove todo list in which the only plan is to make libublksrv as a
single project, however it looks not necessary because now all targets
code has been moved to dedicated directory `targets/`, so remove todo
list entry.

Meantime add one maillist entry for helping to discuss for addressing
issue #96.

Signed-off-by: Ming Lei <tom.leiming@gmail.com>