fix: Handle more interrupts in __ublksrv_ctrl_cmd
Fixes: 8f2a561 ("Handle interrupts in __ublksrv_ctrl_cmd")

The syscall io_uring_enter may return EINTR if it is interrupted by a
delivery of a signal. This can stil occur on the first submit.

This commit moves the previous do-while to a combined
io_uring_submit_and_wait and updates io_uring_wait_cqe to
io_uring_peek_cqe.