info: print the actual flag names for the device flags
in the ublk info output.
This makes it much easier to see for example if and what kind of recovery
settings are active on teh device.

I think some of these flags might be obsolete but lets keep them
in the output for now. If we remove them we should also remove them from
the header file where they are defined.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>