Do not call ublksrv_parse_std_opts() from cmd_dev_recover()
In cmd_dev_recover() we are supposed to parse the options for
"ublk recover ..." but the only one that matters is the
dev_id we get via the -n argument to "ublk recover".

ublksrv_parse_std_opts() is used to prse the arguments to the
"ublk add ..." command and while it also takes a -n argument
for dev_id it is the wrong fucntion to call for this.

Instead create a private helper that only parses the -n argument
and use it isntead.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>