Make ublksrv_ctrl_data a member of ublksrv_ctrl_dev
This structure is only used together with ublksrv_ctrl_dev.
It is used as a stab-based instance with a lifetime
of the device and accessed via private_data.
Instead of this just embed it directly.

Allocate the data structure in ublksrv_ctrl_init and free it
in ublksrv_ctrl_deinit. It now has the same lifetime as
ublksrv_ctrl_dev itself.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>