Make return value from ublksrv_execve_helper consistent.
It used to either return 0 on success of !0 on failure.
!0 either being <0 for explicit checks/return values
and >0 if a call failed setting errno.

Change this to always return <0 on failure.
No change in functionality but it looke more tidy.

Also remove the conditional checking the return code for the call to execve
and returniong an error in the conditional.
Execve will only ever return if there is a failure so having the
conditional is not needed and just confuses to code.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>