targets: remove UBLK_IO_TGT_BUF and UBLK_IO_TGT_IO
Not necessary to add the two flag for deciding if the io_uring OP
belongs to normal UBLK_IO_* io command, and we have carried the exact
ublk op code in ->userdata.

So remove the two.

So far:

	UBLK_IO_OP_* < 0x20

	_IO_NR(UBLK_U_IO_*) > 0x20: register io buffer, unregister io buffer

	UBLK_USER_COPY_READ:  0x80
	UBLK_USER_COPY_WRITE: 0x81

Add is_ublk_io_cmd() helper for checking if the io_uring OP belongs to
normal ublk io command.

In this way, the target data field in ->user_data can be reserved for
future usage, such as the ublk stripe usage:

https://web.git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/tools/testing/selftests/ublk/stripe.c#n148

Signed-off-by: Ming Lei <tom.leiming@gmail.com>