Remove the direct use of ublksrv_print_std_opts from targets
All built-in targets print the target-specific help via the
call to the helper ublksrv_tgt_cmd_main()

This means that a target does not need to print the full set of all
arguments but only those arguments that are unique for that particular
target. Instead we can create the printout of all arguments inside
ublksrv_tgt.cpp:cmd_usage() by combining the standard arguments that
all ublksrv_tgt_cmd_main() consumers support and the extra arguments
unique to this target.

This reduces some code duplication but also removes the need to access
ublksrv_print_std_opts() and thus one less symbol needs to be accessed by
a target.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>