Add options to ublksrv_tgt_type
This field is used with ublksrv_main to detect conditions
where the target uses the same command line argument as
one of the built-in arguments.
When this happens the offending option is printed and
the target will fail to start.

Targets that use ublksrv_main() must provide this field
though this is not yet enforced.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>