Fix valgrind warnings
In several places we request data from the kernel via a buffer.
Valgrind has no way to know that these buffers are actually assigned
and filled in by the kernel and thus will cause warnings to be
triggered as it thinks that the data might be uninitialized.

The warnings are false as the data IS initialized inside the kernel.

Suppress these warnings by explicitely initializing the data by
calling memset (before we pass the pointer to the kernel).

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>