Make ublksrv_tgt_send_dev_event static
We do not need to call this function with devid==-1
if execv fails because we already handle this failure if the
file descriptor is just closed and the parentvprocess fails to
read the dev_id from the pipe.

Thus make it static and reduce the size of the semi-private API in
ublksrv_tgt.cpp.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>