Make ublksrv_main() a public function
ublksrv_main() is the only remaining public symbol that is exported from
targets/ublksrv_tgt.cpp. Move this whole file to lib/ and make this
function part of the public API.

Ublksrv_main() is mostly just boilerplate for creating a target and is
hugely beneficial for both internal as well as out-of-tree targets.
Make it a public function so that all targets can benefit from it and having
to avoid duplicating the code.

As an example, both ublk.iscsi and ublk.nfs that use ublksrv_main(), in both
cases ublksrv_main() and its dependencies is more boilerplate code that the
iscsi/nfs code in ublk.[iscsi|nfs].cpp itself.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>