ublksrv_tgt: move parsing of standard options into the targets
Move the parsing of all the standard options into the targets themselves
via a new ublksrv_tgt_type method .parser_for_add
This allows targets that use ublksrv_main() to pick and choose
what options they want to provide.

It also allows us in later patches to move the target specific options
onto these functions and have a single processing loop for all
commnd line options instead of, as now, one loop for the standard
options for the generic ublk device and a separate loop for the backend
options.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>