Width fix for build_user_data
tgt_data is a 16bit wide field but is being passed as a 32bit wide
parameter which is then shifted left by 24bits causing the upper 8bits
to fall off. First cast this to a 64bit value prior to shifting to
preserve these bits for the usage by the target.