Add an epollfd to the queue's and an API to register fd's and callbacks
This adds an API where ublksrv consumers can register file descriptors
to the internal event-loop that is used to drive a ublksrv queue.

ublksrv_epoll_add_fd() is used to register a new fd to be monitored by
the eventloop and ublksrv_epoll_mod_fd() is used to modify which events
to monitor for.

This API allows to register an application callback that will be
triggered when the monitored events trigger for the file descriptor.
This callback will be invoked from the same eventloop and thread context
that the queue is using to communicate with the ublk_drv kernel module so
it is imperative that the application takes great care that this callback
will never block.

Signed-off-by: Ronnie Sahlberg <rsahlberg@whamcloud.com>
Signed-off-by: Ming Lei <tom.leiming@gmail.com>