Always compile ublk_set_debug_mask and ublk_get_debug_mask
Currently ublk_set_debug_mask and ublk_get_debug_mask are only compiled
if --enable-debug is passed to configure, and are dummied out in
ublksrv_utils.h if DEBUG is not set.

Unfortunately this doesn't work for other programs linking to ublksrv,
as they non-intuitively need to define DEBUG themselves in order for
these functions to be available in the header file.

Fix this by exporting ublk_set_debug_mask and ublk_get_debug_mask
unconditionally, using dummy versions if not in debug mode.

Signed-off-by: Mark Harmstone <mark@harmstone.com>